cleanup: separate helper functions
